requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
fal-client==0.5.6
requests==2.32.3
python-dotenv==1.0.1
